remvoe duplicate pages #111
